refactor: split read renderer/driver/predicate seams (Bundle B)
Pure structural refactor — zero behavior change. Splits two oversized
modules along their natural seams and shares the cap-resolution decision
between the CLI and MCP surfaces.

CLI (src/cli/commands/reads.ts, ~1700 → ~800 lines):
- src/cli/render.ts       — pure renderer family (formatItem, renderList,
                            sidebar/upcoming/logbook/sections, anytime/someday
                            previews, legend, uuidCol/uuidDisplayWidth). No
                            commander dependency.
- src/cli/read-driver.ts  — runRead/withClient envelope driver, parseCap/
                            parseLimit, truncationHint, invocation echo helpers.
- src/cli/period.ts       — parsePeriodStart/End, relativePeriodDate,
                            upcomingBucket, RELATIVE_PERIOD grammar. Fully pure.
- reads.ts keeps command registration + openInThings.

read (src/read/views.ts):
- src/read/sidebar-order.ts — groupBySidebar + rank machinery.
- src/read/predicates.ts    — LIVE/OPEN/ANYTIME_SELF/EFF_PROJECT/
                              PROJECT_ANYTIME_ACTIVE/CONTAINER_UNTRASHED.
- views.ts keeps the view functions + view types.

Typed render boundary:
- withClient is now generic<T>; the list-view table is a generic
  registerListView<T> factory. Fetch return types flow into each renderer's
  parameter — every `render as (d: unknown)` / `(data: never)` cast deleted.
  No mismatch was revealed; the casts were masking correct types.

Shared cap resolution:
- src/read/caps.ts holds the one pure resolveCap(value, all, default) decision.
  Both surfaces keep their own input parsing and error emission.

708 tests pass; typecheck/lint/fmt clean; warning count unchanged (64).
